+++ clang/tools/c-index-test/core_main.cpp
@@ -109,8 +109,8 @@ public:
 
     Loc = SM.getFileLoc(Loc);
     FileID FID = SM.getFileID(Loc);
-    unsigned Line = SM.getLineNumber(FID, SM.getFileOffset(Loc));
-    unsigned Col = SM.getColumnNumber(FID, SM.getFileOffset(Loc));
+    uint64_t Line = SM.getLineNumber(FID, SM.getFileOffset(Loc));
+    uint64_t Col = SM.getColumnNumber(FID, SM.getFileOffset(Loc));
     OS << Line << ':' << Col << " | ";
 
     printSymbolInfo(getSymbolInfo(D), OS);
@@ -147,8 +147,8 @@ public:
 
     Loc = SM.getFileLoc(Loc);
     FileID FID = SM.getFileID(Loc);
-    unsigned Line = SM.getLineNumber(FID, SM.getFileOffset(Loc));
-    unsigned Col = SM.getColumnNumber(FID, SM.getFileOffset(Loc));
+    uint64_t Line = SM.getLineNumber(FID, SM.getFileOffset(Loc));
+    uint64_t Col = SM.getColumnNumber(FID, SM.getFileOffset(Loc));
     OS << Line << ':' << Col << " | ";
 
     printSymbolInfo(getSymbolInfo(ImportD), OS);
@@ -170,8 +170,8 @@ public:
 
     Loc = SM.getFileLoc(Loc);
     FileID FID = SM.getFileID(Loc);
-    unsigned Line = SM.getLineNumber(FID, SM.getFileOffset(Loc));
-    unsigned Col = SM.getColumnNumber(FID, SM.getFileOffset(Loc));
+    uint64_t Line = SM.getLineNumber(FID, SM.getFileOffset(Loc));
+    uint64_t Col = SM.getColumnNumber(FID, SM.getFileOffset(Loc));
     OS << Line << ':' << Col << " | ";
 
     printSymbolInfo(getSymbolInfoForMacro(*MI), OS);
